requirements.txt
torch>=2.0.1
transformers>=4.38.2
datasets>=2.16.1
peft>=0.9.0
safetensors>=0.4.2
accelerate>=0.27.2
python-dotenv
trl
